+++ pkgs/ffigen/lib/src/code_generator/objc_protocol.dart
@@ -176,14 +176,15 @@ interface class $name extends $protocolBase $impls{
 ''');
       }
 
-      final args = buildArgs.isEmpty ? '' : '{${buildArgs.join(', ')}}';
+      buildArgs.add('bool \$keepIsolateAlive = false');
+      final args = '{${buildArgs.join(', ')}}';
       final builders = '''
   /// Builds an object that implements the $originalName protocol. To implement
   /// multiple protocols, use [addToBuilder] or [$protocolBuilder] directly.
   static $name implement($args) {
     final builder = $protocolBuilder();
     $buildImplementations
-    return $name.castFrom(builder.build());
+    return $name.castFrom(builder.build(keepIsolateAlive: \$keepIsolateAlive));
   }
 
   /// Adds the implementation of the $originalName protocol to an existing
@@ -202,7 +203,7 @@ interface class $name extends $protocolBase $impls{
   static $name implementAsListener($args) {
     final builder = $protocolBuilder();
     $buildListenerImplementations
-    return $name.castFrom(builder.build());
+    return $name.castFrom(builder.build(keepIsolateAlive: \$keepIsolateAlive));
   }
 
   /// Adds the implementation of the $originalName protocol to an existing
@@ -218,7 +219,7 @@ interface class $name extends $protocolBase $impls{
   static $name implementAsBlocking($args) {
     final builder = $protocolBuilder();
     $buildBlockingImplementations
-    return $name.castFrom(builder.build());
+    return $name.castFrom(builder.build(keepIsolateAlive: \$keepIsolateAlive));
   }
 
   /// Adds the implementation of the $originalName protocol to an existing